Add steps from email steps to send invite steps Delete unnecessary email steps
